chore: remove useFilteredPinnedAssets hook and use in-line instead
